+++ tests/image/test_image_gradients.py
