+++ torchmetrics/functional/image/__init__.py
@@ -13,3 +13,4 @@
 # limitations under the License.
 
 from torchmetrics.functional.image.psnr import psnr  # noqa: F401
+from torchmetrics.functional.image.image_gradients import image_gradients  # noqa: F401
